pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-auth-os"
version = "0.1.0"
description = "Secure authentication system with MFA, RBAC, session management, and OS integration abstraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "SQLAlchemy>=2.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "passlib[argon2]>=1.7.4",
  "argon2-cffi>=23.1.0",
  "pyotp>=2.9.0",
  "qrcode[pil]>=7.4.2",
  "cryptography>=42.0.0",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.2.0",
  "typer>=0.12.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
packages = ["app"]
